Migliorate funzionalità della user page, con nuovi metodi di aggiunta ed eliminazione dei  pet

+++ frontend/src/views/UserPage.vue
@@ -13,15 +13,40 @@
           <b>Name:</b> {{user[0].name}} <br>
           <b>Username:</b> {{user[0].username}} <br>
           <b>Mail address:</b> {{user[0].mail}} <br>
-          <b>List of pets:</b> {{pets}}
+          <b>List of pets:</b>
+            <ul v-for="pet in pets" style="list-style: none;">
+              <li>
+                <div class="card mb-3">
+                  <div class="card-header text-end">
+                    <a type="button" class="btn-close" aria-label="Close" @click="removePets(pet._id)"></a>
+                  </div>
+                  <div class="row g-0">
+                    <div class="col-md-4">
+                      <img :src="pet.pictures[0]" class="img-fluid rounded-start" alt="" style="max-height: 100px; object-fit:cover;"/>
+                    </div>
+                    <div class="col-md-8">
+                      <div class="card-body">
+                        <h5 class="card-title">Name: {{ pet.name }}</h5>
+                        <p class="card-text">Race: {{ pet.race }}</p>
+                        <p class="card-text">Age: {{ pet.age }}</p>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </li>
+            </ul>
         </p>
         <label for="petname">Name of the pet to add</label>
         <input id="petname" type="text" class="form-control form-control-lg" />
-        <button type="button" class="btn btn-dark btn-lg btn-block" id="addpets" @click="addPets()">Add a new pet</button>
+        <label for="petrace">Race of the pet to add</label>
+        <input id="petrace" type="text" class="form-control form-control-lg" />
+        <label for="petage">Age of the pet to add</label>
+        <input id="petage" type="text" class="form-control form-control-lg" />
+        <label for="petimg">Image of the pet to add</label>
+        <input id="petimg" type="file" class="form-control form-control-lg" multiple/>
         <br>
-        <label for="petname2">Name of the pet to remove</label>
-        <input id="petname2" type="text" class="form-control form-control-lg" />
-        <button type="button" class="btn btn-dark btn-lg btn-block" id="addpets" @click="removePets()">Remove the pet</button>
+        <button type="button" class="btn btn-dark btn-lg btn-block" id="addpets" @click="addPets()">Add a new pet</button>
+
       </div>
     </div>
   </section>
@@ -31,6 +56,7 @@
 <script>
 import axios from 'axios'
 import SiteFooterVue from '@/components/SiteFooter.vue'
+import Swal from 'sweetalert2'
 
 export default {
   name: 'UserPage',
@@ -42,8 +68,10 @@ export default {
       })
     axios.get('/api/users/userPets')
       .then((response) => {
-        this.pets.push(response.data)
-        console.log(this.pets.length)
+        for (let i = 0; i < response.data.length; i++) {
+          this.pets.push(response.data[i])
+          console.log(this.pets.length)
+        }
       })
   },
   components: { SiteFooterVue },
@@ -56,9 +84,10 @@ export default {
   methods: {
     addPets: async function () {
       const pet = {
-        name: document.getElementById('petname').value
+        name: document.getElementById('petname').value,
+        race: document.getElementById('petrace').value,
+        age: document.getElementById('petage').value
       }
-      console.log(document.getElementById('petname').value)
       fetch('/api/pets', {
         method: 'post',
         headers: {
@@ -66,13 +95,65 @@ export default {
         },
         body: JSON.stringify(pet)
       })
+        .then((response) => response.json())
+        .then((data) => {
+          const id = data.ownedAnimals[data.ownedAnimals.length - 1]
+          const formData = new FormData()
+          const images = document.querySelector('#petimg')
+          for (let i = 0; i <= images.files.length; i++) {
+            formData.append('images', images.files[i])
+          }
+          axios
+            .post(
+              '/api/images/pets',
+              formData,
+              {
+                params: {
+                  id: id
+                }
+              },
+              {
+                headers: {
+                  'Content-Type': 'multipart/form-data'
+                }
+              }
+            )
+        })
+        .then(() => {
+          Swal.fire({
+            icon: 'success',
+            title: 'You added the pet successfully!',
+            showConfirmButton: true
+          })
+            .then((response) => {
+              if (response.isConfirmed) location.reload()
+            })
+        })
     },
-    removePets: async function () {
-      console.log(document.getElementById('petname').value)
-      fetch('/api/pets?pet=' + document.getElementById('petname2').value, {
-        method: 'delete',
-        headers: {
-          'Content-type': 'application/json'
+    removePets: async function (id) {
+      Swal.fire({
+        title: 'Do you want to remove this pet from the list?',
+        showCancelButton: true,
+        confirmButtonText: 'Yes'
+      }).then((res) => {
+        if (res.isConfirmed) {
+          console.log(document.getElementById('petname').value)
+          fetch('/api/pets?pet=' + id, {
+            method: 'delete',
+            headers: {
+              'Content-type': 'application/json'
+            }
+          })
+            .then(() => {
+              Swal.fire({
+                icon: 'success',
+                title: 'You removed the pet successfully!',
+                showConfirmButton: true
+              })
+                .then((response) => {
+                  if (response.isConfirmed) location.reload()
+                })
+            })
         }
       })
     }
